+++ master-firmware/package.yml
@@ -46,8 +46,9 @@ target.arm:
     - src/aversive_port/cvra_motors.c
     - src/aversive_port/cvra_pid.c
     - src/base/encoder.c
-    - src/gui.c
-    - src/gui_utilities.c
+    - src/gui.cpp
+    - src/gui/Menu.cpp
+    - src/gui/MenuPage.cpp
     - src/base/base_controller.c
     - src/base/map_server.cpp
     - src/arms/cvra_arm_motors.c
@@ -67,8 +68,6 @@ target.arm:
     - src/strategy/color_sequence_server.c
     - src/ballgun/ballgun_module.c
     - src/ballgun/ball_sense.c
-    - src/gui/pages/root.c
-    - src/gui/pages/position.c
 
 source:
     - src/unix_timestamp.c
@@ -115,7 +114,6 @@ source:
     - src/msgbus_protobuf.c
     - src/ballgun/ballgun.c
     - src/ballgun/ball_sensor.c
-    - src/gui/menu.c
 
 include_directories:
     - src/
