+++ master-firmware/src/gui.h
@@ -4,6 +4,7 @@
 #include <error/error.h>
 #include <stdarg.h>
 
+
 #ifdef __cplusplus
 extern "C" {
 #endif
