--- master-firmware/src/gui_utilities.h
@@ -1,3 +1,0 @@
-void page_1(void);
-void page_3(void);
-void page_2(void);
