+++ ugfx_demo/gfxconf.h
@@ -73,6 +73,7 @@
 
 #define GWIN_REDRAW_IMMEDIATE GFXON
 
+
 ///////////////////////////////////////////////////////////////////////////
 // GEVENT                                                                //
 ///////////////////////////////////////////////////////////////////////////
